pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "instatrace"
version = "1.0.0"
description = "Ultimate Instagram OSINT Intelligence Tool"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.8"
authors = [
    {name = "John-Varghese-EH", email = ""}
]
keywords = ["instagram", "osint", "recon", "intelligence", "security", "research"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet",
]
dependencies = [
    "requests>=2.28.0",
    "colorama>=0.4.6",
    "beautifulsoup4>=4.12.0",
    "phonenumbers>=8.13.0",
    "pycountry>=22.3.0",
    "python-whois>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/John-Varghese-EH/InstaTrace"
Repository = "https://github.com/John-Varghese-EH/InstaTrace"
Issues = "https://github.com/John-Varghese-EH/InstaTrace/issues"

[project.scripts]
instatrace = "instatrace.cli:main"

[tool.setuptools.packages.find]
include = ["instatrace*"]
